Rename secrets to get-secret
